% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.package.R
\docType{package}
\name{fastCNVdata-package}
\alias{fastCNVdata}
\alias{fastCNVdata-package}
\title{fastCNVdata: Data package for fastCNV containing single cell and spatial transcriptomics data}
\description{
This package contains single cell and spatial transcriptomics data from different tissues
}
\author{
\strong{Maintainer}: Gadea Cabrejas \email{gadea.cabrejas-saiz@u-paris.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{##}
\keyword{(with}
\keyword{:}
\keyword{annotations)}
\keyword{data}
\keyword{from}
\keyword{https://doi.org/10.1038/s41698-023-00488-4}
\keyword{internal}
\keyword{new}
\keyword{paper}
\keyword{valdeolivas}
