% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDBreast.R
\docType{data}
\name{HDBreast}
\alias{HDBreast}
\title{Spatial transcriptomics HD data from Human breast cancer}
\format{
An object of class Seurat.
}
\source{
10X Genomics datasets
}
\usage{
data(HDBreast)
}
\description{
Data from https://www.10xgenomics.com/datasets
}
\keyword{datasets}
