% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scBreast.R
\docType{data}
\name{scBreast}
\alias{scBreast}
\title{scRNA-seq data from Human brest cancer}
\format{
An object of class Seurat.
}
\source{
https://www.10xgenomics.com/datasets

https://www.10xgenomics.com/datasets/7-5-k-sorted-cells-from-human-invasive-ductal-carcinoma-3-v-3-1-3-1-standard-6-0-0
}
\usage{
data(scBreast)
}
\description{
Data downloaded from the 10X genomics data base
}
\keyword{datasets}
