% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scColon.R
\docType{data}
\name{scColon}
\alias{scColon}
\title{scRNA-seq data from Human colorectal cancer}
\format{
An object of class Seurat.
}
\source{
https://pubmed.ncbi.nlm.nih.gov/35773407/
}
\usage{
data(scColon)
}
\description{
Data from : Single-cell and bulk transcriptome sequencing identifies two epithelial
tumor cell states and refines the consensus molecular classification of colorectal cancer.
Nat Genet. 2022 Jul;54(7):963-975. doi: 10.1038/s41588-022-01100-4.
}
\keyword{datasets}
