% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scColon3.R
\docType{data}
\name{scColon3}
\alias{scColon3}
\title{scRNA-seq data from Human colorectal cancer}
\format{
An object of class Seurat.
}
\source{
https://pubmed.ncbi.nlm.nih.gov/34450029/
}
\usage{
data(scColon3)
}
\description{
Data from :  Spatially organized multicellular immune hubs in human colorectal
cancer. Cell. 2021 Sep 2;184(18):4734-4752.e20. doi: 10.1016/j.cell.2021.08.003.
Epub 2021 Aug 26. PMID: 34450029; PMCID: PMC8772395.
}
\keyword{datasets}
