% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STColon2.R
\docType{data}
\name{STColon2}
\alias{STColon2}
\title{Spatial transcriptomics data from Human colorectal cancer}
\format{
An object of class Seurat.
}
\source{
Valdeolivas, A., Amberg, B., Giroud, N. et al. Profiling the heterogeneity of colorectal cancer consensus molecular subtypes using spatial transcriptomics. npj Precis. Onc. 8, 10 (2024). https://doi.org/10.1038/s41698-023-00488-4
}
\usage{
data(STColon2)
}
\description{
Data from https://doi.org/10.1038/s41698-023-00488-4
}
\keyword{datasets}
